fix(ci): restore main validation lanes
