+++ src/Meridian.Ui.Shared/Endpoints/FundStructureEndpoints.cs
@@ -1,4 +1,3 @@
-using System.Text;
 using System.Text.Json;
 using Meridian.Identity.Auth;
 using Meridian.Application.FundStructure;
@@ -8,7 +7,6 @@ using Meridian.Contracts.Services;
 using Meridian.Contracts.Workstation;
 using Meridian.Entities.FundStructure;
 using Meridian.Reporting;
-using Meridian.Storage.Export;
 using Meridian.Ui.Shared.Services;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
@@ -2950,332 +2948,6 @@ public static partial class FundStructureEndpoints
             error = $"assignmentReference is invalid for '{LedgerGroupingRules.LedgerGroupAssignmentType}'. {LedgerGroupId.ValidationMessage}";
             return request;
         }
-    }
-
-    private const string StructuredXlsxContentType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
-
-    private static bool IsStructuredCsvRequest(string? format) =>
-        string.Equals(format, "csv", StringComparison.OrdinalIgnoreCase);
-
-    private static bool IsStructuredJsonRequest(string? format) =>
-        string.Equals(format, "json", StringComparison.OrdinalIgnoreCase);
-
-    private static bool IsPortalJsonRequest(string? format) =>
-        string.Equals(format, "json", StringComparison.OrdinalIgnoreCase);
-
-    private static bool IsStructuredXlsxRequest(string? format) =>
-        string.Equals(format, "xlsx", StringComparison.OrdinalIgnoreCase)
-        || string.Equals(format, "xls", StringComparison.OrdinalIgnoreCase)
-        || string.Equals(format, "excel", StringComparison.OrdinalIgnoreCase);
-
-    private static string BuildDeliveryPortalHtml(ReportPackDeliveryPackageDto package)
-    {
-        var theme = package.BrandingTheme;
-        var primaryColor = NormalizePortalColor(theme?.PrimaryColor, "#1F4E79");
-        var accentColor = NormalizePortalColor(theme?.AccentColor, "#2F9E8F");
-        var textColor = NormalizePortalColor(theme?.TextColor, "#111827");
-        var backgroundColor = NormalizePortalColor(theme?.BackgroundColor, "#FFFFFF");
-        var firmName = string.IsNullOrWhiteSpace(theme?.FirmName)
-            ? "Meridian"
-            : theme.FirmName.Trim();
-        var builder = new StringBuilder();
-        builder.AppendLine("<!doctype html>");
-        builder.Append("<html lang=\"en\"><head><meta charset=\"utf-8\"><meta name=\"viewport\" content=\"width=device-width,initial-scale=1\">")
-            .Append("<title>")
-            .Append(EscapePortalHtml(firmName))
-            .Append(" Report Package</title><style>");
-        builder.Append("body{margin:0;font-family:Arial,sans-serif;background:")
-            .Append(backgroundColor)
-            .Append(";color:")
-            .Append(textColor)
-            .Append("}.portal-header{border-top:8px solid ")
-            .Append(primaryColor)
-            .Append(";padding:28px 32px 18px}.portal-kicker{color:")
-            .Append(accentColor)
-            .Append(";font-size:12px;text-transform:uppercase;letter-spacing:.12em}.content{padding:0 32px 32px;max-width:980px}.grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(220px,1fr));gap:12px}.panel{border:1px solid #d0d5dd;border-radius:6px;padding:14px;background:rgba(255,255,255,.72)}a{color:")
-            .Append(primaryColor)
-            .Append("}.artifact{display:flex;gap:8px;align-items:flex-start;justify-content:space-between;border-top:1px solid #eaecf0;padding:10px 0}.badge{border:1px solid ")
-            .Append(accentColor)
-            .Append(";border-radius:999px;padding:2px 8px;font-size:11px;text-transform:uppercase}.footer{border-top:2px solid ")
-            .Append(accentColor)
-            .Append(";margin-top:24px;padding-top:14px;font-size:12px}</style></head><body>");
-        builder.Append("<header class=\"portal-header\"><div class=\"portal-kicker\">")
-            .Append(EscapePortalHtml(firmName))
-            .Append("</div><h1>Secure Report Package</h1><p>")
-            .Append(EscapePortalHtml(package.DeliveryAccessSummary ?? "Token-gated report package."))
-            .Append("</p>");
-        if (!string.IsNullOrWhiteSpace(theme?.LogoUri))
-        {
-            builder.Append("<p>Logo: ")
-                .Append(EscapePortalHtml(theme.LogoUri.Trim()))
-                .Append("</p>");
-        }
-
-        builder.AppendLine("</header><main class=\"content\">");
-        builder.AppendLine("<section class=\"grid\" aria-label=\"Package details\">");
-        AppendPortalSummaryPanel(builder, "Package", package.PackageId, package.DeliveryChannelSummary ?? package.DeliveryMode.ToString());
-        AppendPortalSummaryPanel(builder, "Integrity", package.IntegritySummary ?? "Integrity summary not retained.", package.RetainedManifestPath);
-        AppendPortalSummaryPanel(builder, "Access", package.AccessExpiresAtUtc?.ToString("O") ?? "No expiry retained.", package.PortalRoute);
-        AppendPortalSummaryPanel(builder, "Dataset", FormatPortalDatasetSummary(package), package.ReportingRunId ?? package.PublicationManifestId ?? "Publication package");
-        builder.AppendLine("</section>");
-        builder.AppendLine("<section class=\"panel\" aria-label=\"Package artifacts\"><h2>Downloads</h2>");
-        foreach (var artifact in package.Artifacts ?? [])
-        {
-            builder.Append("<div class=\"artifact\"><span><strong>")
-                .Append(EscapePortalHtml(artifact.ArtifactName))
-                .Append("</strong><br><span>")
-                .Append(EscapePortalHtml(artifact.ContentType))
-                .Append("</span><br><span>SHA-256 ")
-                .Append(EscapePortalHtml(artifact.ChecksumSha256))
-                .Append("</span></span><span><span class=\"badge\">")
-                .Append(EscapePortalHtml(artifact.Format.ToString()))
-                .Append("</span> ");
-            if (!string.IsNullOrWhiteSpace(artifact.DownloadRoute))
-            {
-                builder.Append("<a href=\"")
-                    .Append(EscapePortalAttribute(artifact.DownloadRoute))
-                    .Append("\">Download</a>");
-            }
-
-            builder.AppendLine("</span></div>");
-        }
-
-        builder.AppendLine("</section>");
-        if (package.Notifications is { Count: > 0 })
-        {
-            builder.AppendLine("<section class=\"panel\" aria-label=\"Package notifications\"><h2>Notifications</h2><ul>");
-            foreach (var notification in package.Notifications)
-            {
-                builder.Append("<li><strong>")
-                    .Append(EscapePortalHtml(notification.Subject))
-                    .Append("</strong><br>")
-                    .Append(EscapePortalHtml(notification.Body))
-                    .Append("</li>");
-            }
-
-            builder.AppendLine("</ul></section>");
-        }
-
-        if (theme is not null)
-        {
-            builder.Append("<footer class=\"footer\"><strong>")
-                .Append(EscapePortalHtml(theme.Name))
-                .Append("</strong><br>")
-                .Append(EscapePortalHtml(theme.FooterText ?? string.Empty));
-            if (!string.IsNullOrWhiteSpace(theme.Disclaimer))
-            {
-                builder.Append("<br>")
-                    .Append(EscapePortalHtml(theme.Disclaimer.Trim()));
-            }
-
-            builder.AppendLine("</footer>");
-        }
-
-        builder.AppendLine("</main></body></html>");
-        return builder.ToString();
-    }
-
-    private static void AppendPortalSummaryPanel(
-        StringBuilder builder,
-        string label,
-        string value,
-        string detail)
-    {
-        builder.Append("<article class=\"panel\"><h2>")
-            .Append(EscapePortalHtml(label))
-            .Append("</h2><p><strong>")
-            .Append(EscapePortalHtml(value))
-            .Append("</strong></p><p>")
-            .Append(EscapePortalHtml(detail))
-            .AppendLine("</p></article>");
-    }
-
-    private static string FormatPortalDatasetSummary(ReportPackDeliveryPackageDto package)
-    {
-        var source = package.ReportWriterDatasetSourceLabel
-            ?? package.ReportWriterDatasetSourceId
-            ?? "Report package";
-        return package.ReportWriterDatasetRowCount.HasValue
-            ? $"{source} ({package.ReportWriterDatasetRowCount.Value} rows)"
-            : source;
-    }
-
-    private static string NormalizePortalColor(string? value, string fallback)
-    {
-        if (string.IsNullOrWhiteSpace(value))
-        {
-            return fallback;
-        }
-
-        var normalized = value.Trim();
-        return normalized.Length == 7
-               && normalized[0] == '#'
-               && normalized.Skip(1).All(Uri.IsHexDigit)
-            ? normalized
-            : fallback;
-    }
-
-    private static string EscapePortalHtml(string? value) =>
-        (value ?? string.Empty)
-            .Replace("&", "&amp;", StringComparison.Ordinal)
-            .Replace("<", "&lt;", StringComparison.Ordinal)
-            .Replace(">", "&gt;", StringComparison.Ordinal)
-            .Replace("\"", "&quot;", StringComparison.Ordinal);
-
-    private static string EscapePortalAttribute(string? value) =>
-        EscapePortalHtml(value).Replace("'", "&#39;", StringComparison.Ordinal);
-
-    private static byte[] BuildStructuredExportJson(
-        StructuredReportingExportPayloadDto payload,
-        JsonSerializerOptions jsonOptions) =>
-        JsonSerializer.SerializeToUtf8Bytes(payload, jsonOptions);
-
-    private static byte[] BuildStructuredExportCsv(StructuredReportingExportPayloadDto payload)
-    {
-        var builder = new StringBuilder();
-        AppendStructuredExportCsvRow(builder, payload.Columns.Select(static column => column.Name));
-
-        foreach (var row in payload.Rows)
-        {
-            AppendStructuredExportCsvRow(
-                builder,
-                payload.Columns.Select(column =>
-                    row.TryGetValue(column.Name, out var value) ? value : null));
-        }
-
-        return Encoding.UTF8.GetBytes(builder.ToString());
-    }
-
-    private static byte[] BuildStructuredExportXlsx(StructuredReportingExportPayloadDto payload)
-    {
-        var headers = payload.Columns.Select(static column => column.Name).ToArray();
-        var rows = payload.Rows
-            .Select(row => headers
-                .Select(header => row.TryGetValue(header, out var value) ? value : null)
-                .Cast<object?>()
-                .ToArray())
-            .Cast<IReadOnlyList<object?>>()
-            .ToArray();
-
-        return XlsxWorkbookWriter.CreateWorkbook(
-        [
-            new XlsxWorksheet(payload.Export.Dataset, headers, rows),
-            new XlsxWorksheet(
-                "Metadata",
-                ["Field", "Value"],
-                BuildStructuredExportMetadataRows(payload)),
-            new XlsxWorksheet(
-                "DataDictionary",
-                ["Ordinal", "Name", "Data type", "Required", "Description"],
-                BuildStructuredExportDataDictionaryRows(payload)),
-            new XlsxWorksheet(
-                "Validation",
-                ["Check", "Status", "Detail"],
-                BuildStructuredExportValidationRows(payload)),
-            new XlsxWorksheet(
-                "RowLineage",
-                ["Row", "Row key", "SHA-256"],
-                BuildStructuredExportRowLineageRows(payload))
-        ]);
-    }
-
-    private static IReadOnlyList<IReadOnlyList<object?>> BuildStructuredExportMetadataRows(
-        StructuredReportingExportPayloadDto payload) =>
-    [
-        ["exportId", payload.Export.ExportId],
-        ["label", payload.Export.Label],
-        ["purpose", payload.Export.Purpose.ToString()],
-        ["dataset", payload.Export.Dataset],
-        ["consumer", payload.Export.Consumer],
-        ["schemaVersion", payload.Export.SchemaVersion],
-        ["rowCount", payload.Export.RowCount],
-        ["fieldCount", payload.Export.FieldCount],
-        ["sourceCount", payload.Export.SourceCount],
-        ["currency", payload.Export.Currency],
-        ["asOf", payload.Export.AsOf],
-        ["isReady", payload.Export.IsReady],
-        ["retainedPath", payload.Export.RetainedPath],
-        ["route", payload.Export.Route],
-        ["versionStamp", payload.Export.VersionStamp],
-        ["generatedAtUtc", payload.GeneratedAtUtc],
-        ["generatedByPrincipalId", payload.GeneratedByPrincipalId],
-        ["generatedForCompanyId", payload.GeneratedForCompanyId],
-        ["generatedForGroups", string.Join(";", payload.GeneratedForGroupPrincipalIds ?? [])],
-        ["rowLineageCount", payload.RowLineage?.Count ?? 0]
-    ];
-
-    private static IReadOnlyList<IReadOnlyList<object?>> BuildStructuredExportDataDictionaryRows(
-        StructuredReportingExportPayloadDto payload)
-    {
-        var fields = payload.DataDictionary is { Count: > 0 }
-            ? payload.DataDictionary
-            : payload.Columns.Select((column, index) => new StructuredReportingExportDataDictionaryFieldDto(
-                column.Name,
-                column.DataType,
-                column.Description,
-                index + 1)).ToArray();
-
-        return fields
-            .Select(static field => (IReadOnlyList<object?>)
-            [
-                field.Ordinal,
-                field.Name,
-                field.DataType,
-                field.Required ? "true" : "false",
-                field.Description
-            ])
-            .ToArray();
-    }
-
-    private static IReadOnlyList<IReadOnlyList<object?>> BuildStructuredExportValidationRows(
-        StructuredReportingExportPayloadDto payload) =>
-        (payload.ValidationChecks ?? [])
-            .Select(static check => (IReadOnlyList<object?>)
-            [
-                check.CheckId,
-                check.Status,
-                check.Detail
-            ])
-            .ToArray();
-
-    private static IReadOnlyList<IReadOnlyList<object?>> BuildStructuredExportRowLineageRows(
-        StructuredReportingExportPayloadDto payload) =>
-        (payload.RowLineage ?? [])
-            .Select(static row => (IReadOnlyList<object?>)
-            [
-                row.RowNumber,
-                row.RowKey,
-                row.RowHashSha256
-            ])
-            .ToArray();
-
-    private static void AppendStructuredExportCsvRow(StringBuilder builder, IEnumerable<string?> values)
-    {
-        var first = true;
-        foreach (var value in values)
-        {
-            if (!first)
-            {
-                builder.Append(',');
-            }
-
-            builder.Append(EscapeStructuredExportCsvValue(value));
-            first = false;
-        }
-
-        builder.AppendLine();
-    }
-
-    private static string EscapeStructuredExportCsvValue(string? value)
-    {
-        if (string.IsNullOrEmpty(value))
-        {
-            return string.Empty;
-        }
-
-        return value.IndexOfAny([',', '"', '\r', '\n']) >= 0
-            ? $"\"{value.Replace("\"", "\"\"", StringComparison.Ordinal)}\""
-            : value;
     }
 
     private static int ParseInt(string? value, int defaultValue) =>
