+++ src/Meridian.Ui.Shared/Services/ReportPackDeliveryService.Guards.cs
@@ -1,0 +1,15 @@
+using Meridian.Contracts.Workstation;
+
+namespace Meridian.Ui.Shared.Services;
+
+public sealed partial class ReportPackDeliveryService
+{
+    private static void EnsureHumanOrigin(OperationsActionOriginDto actionOrigin, string action)
+    {
+        if (actionOrigin != OperationsActionOriginDto.HumanOperator)
+        {
+            throw new InvalidOperationException(
+                $"Reviewed automation cannot {action}; a human operator approval is required.");
+        }
+    }
+}
