+++ src/Meridian.Ui/dashboard/src/lib/api/reporting-runs.api.ts
@@ -1,0 +1,16 @@
+import type { ApiRequestOptions } from "@/lib/api";
+import { FUND_STRUCTURE_API_ENDPOINTS } from "@/lib/ui-api-routes.generated";
+import type { ReportingRunReadiness, ReportingRunRequest, ReportingRunResult } from "@/types";
+
+type PostJson = <T>(path: string, body?: unknown, options?: ApiRequestOptions) => Promise<T>;
+
+export function createReportingRunsApi(postJson: PostJson) {
+  return {
+    runReportingNow(request: ReportingRunRequest, options: ApiRequestOptions = {}) {
+      return postJson<ReportingRunResult>(FUND_STRUCTURE_API_ENDPOINTS.reportingRuns, request, options);
+    },
+    assessReportingRunReadiness(request: ReportingRunRequest, options: ApiRequestOptions = {}) {
+      return postJson<ReportingRunReadiness>(FUND_STRUCTURE_API_ENDPOINTS.reportingRunReadiness, request, options);
+    }
+  };
+}
