+++ src/Meridian.Ui/dashboard/src/lib/ui-api-routes.generated.ts
@@ -758,6 +758,13 @@ export const UI_API_ROUTES = {
   ReportingStructuredExport: "/api/fund-structure/reporting/structured-exports/{exportId}",
   ReportingRuns: "/api/fund-structure/reporting/runs",
   ReportingRunReadiness: "/api/fund-structure/reporting/runs/readiness",
+  ReportingGovernedRun: "/api/fund-structure/reporting/runs/{runId}",
+  ReportingGovernedRunValidate: "/api/fund-structure/reporting/runs/{runId}/validate",
+  ReportingGovernedRunSubmit: "/api/fund-structure/reporting/runs/{runId}/submit",
+  ReportingGovernedRunApprove: "/api/fund-structure/reporting/runs/{runId}/approve",
+  ReportingGovernedRunRelease: "/api/fund-structure/reporting/runs/{runId}/release",
+  ReportingGovernedRunRestatementRequests: "/api/fund-structure/reporting/runs/{runId}/restatement-requests",
+  ReportingGovernedRestatementApprove: "/api/fund-structure/reporting/runs/restatement-requests/{requestId}/approve",
   ReportingRunAuditTrail: "/api/fund-structure/reporting/runs/{runId}/audit",
   ReportingRunStream: "/api/fund-structure/reporting/runs/{runId}/stream",
   ReportingRunReportWriterGrid: "/api/fund-structure/reporting/runs/{runId}/report-writer-grids/{gridId}",
