+++ src/Meridian.Ui/dashboard/src/screens/data-screen.view-model.test.ts
@@ -481,6 +481,8 @@ describe("data-screen view model", () => {
     expect(resolveDataWorkstream("/data/backfills")).toBe("backfills");
     expect(resolveDataWorkstream("/data/providers")).toBe("providers");
     expect(resolveDataWorkstream("/data/import")).toBe("import");
+    expect(resolveDataWorkstream("/data/operations")).toBe("operations");
+    expect(resolveDataWorkstream("/data/assurance")).toBe("assurance");
     expect(resolveDataWorkstream("/data/exports")).toBe("exports");
     expect(resolveDataWorkstream("/data/query")).toBe("query");
     expect(resolveDataWorkstream("/data")).toBe("overview");
