+++ tests/scripts/test_release_evidence_workflows.py
@@ -7,6 +7,7 @@ from pathlib import Path
 REPO_ROOT = Path(__file__).resolve().parents[2]
 PUBLISH_SMOKE = REPO_ROOT / ".github" / "workflows" / "publish-smoke.yml"
 DESKTOP_INSTALLER = REPO_ROOT / ".github" / "workflows" / "desktop-installer-packaging.yml"
+ROBINHOOD_OPTIONS_SMOKE = REPO_ROOT / ".github" / "workflows" / "robinhood-options-smoke.yml"
 
 
 class ReleaseEvidenceWorkflowTests(unittest.TestCase):
@@ -25,6 +26,14 @@ class ReleaseEvidenceWorkflowTests(unittest.TestCase):
         self.assertIn("--project desktop-installer", workflow)
         self.assertIn("--output artifacts/release/${{ matrix.runtime }}/release-evidence.json", workflow)
         self.assertIn("artifacts/release/**/release-evidence.json", workflow)
+
+    def test_robinhood_smoke_uses_named_powershell_splatting(self) -> None:
+        workflow = ROBINHOOD_OPTIONS_SMOKE.read_text(encoding="utf-8")
+
+        self.assertIn("$smokeArgs = @{", workflow)
+        self.assertIn("Configuration = '${{ inputs.configuration }}'", workflow)
+        self.assertIn("$smokeArgs.SkipBuild = $true", workflow)
+        self.assertNotIn("@('-Configuration'", workflow)
 
 
 if __name__ == "__main__":
