+++ tests/scripts/test_run_provider_validation_evidence_bundle.py
@@ -19,6 +19,48 @@ SCRIPT_PATH = REPO_ROOT / "scripts" / "dev" / "run-provider-validation-evidence-
 
 
 class RunProviderValidationEvidenceBundleTests(unittest.TestCase):
+    def test_prepare_operator_signoff_succeeds_without_claiming_dk1_exit(self) -> None:
+        pwsh = shutil.which("pwsh")
+        if pwsh is None:
+            self.skipTest("pwsh is not available")
+
+        with tempfile.TemporaryDirectory() as temp_dir:
+            root = Path(temp_dir)
+            output_root = root / "bundle-output"
+            date_stamp = "2026-05-01"
+            bundle_root = output_root / date_stamp
+            bundle_root.mkdir(parents=True, exist_ok=True)
+
+            summary_path = bundle_root / "wave1-validation-summary.json"
+            summary = _build_passing_summary()
+            summary["generatedAtUtc"] = "2026-05-01T10:00:00Z"
+            summary_path.write_text(json.dumps(summary), encoding="utf-8")
+
+            completed = subprocess.run(
+                [
+                    pwsh,
+                    "-NoProfile",
+                    "-File",
+                    str(SCRIPT_PATH),
+                    "-DateStamp",
+                    date_stamp,
+                    "-OutputRoot",
+                    str(output_root),
+                    "-SkipWave1",
+                    "-PrepareOperatorSignoff",
+                ],
+                cwd=REPO_ROOT,
+                check=True,
+                capture_output=True,
+                text=True,
+            )
+
+            self.assertIn("ready for operator review", completed.stdout)
+            signoff = json.loads((bundle_root / "dk1-operator-signoff.json").read_text(encoding="utf-8-sig"))
+            self.assertTrue(signoff["packetReview"]["validForOperatorReview"])
+            self.assertTrue(all(approval["decision"] == "pending" for approval in signoff["approvals"]))
+            self.assertFalse((bundle_root / "provider-validation-evidence-bundle.json").exists())
+
     def test_skip_wave1_generates_traceable_bundle_with_expected_artifact_paths(self) -> None:
         pwsh = shutil.which("pwsh")
         if pwsh is None:
